move `DocumentHead` call and SEO content into isolated component
